+++ packages/fighting-components/alert/src/interface.d.ts
@@ -3,3 +3,10 @@ export type alertType = 'primary' | 'success' | 'danger' | 'warning' | 'default'
 export interface handleCloseInterface {
   (evt: MouseEvent): void
 }
+
+export interface fixedStyleInterface {
+  top: ?string
+  button: ?string
+  right: ?string
+  left: ?string
+}